+++ CHANGELOG.md
@@ -38,3 +38,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
 ### Improvements
+
+* [#24](https://github.com/cosmos/cosmos-sdk/pull/24) Validate public randomness commitment.